requirements.txt
streamlit
speechRecognition
pyttsx3
pymysql
python-dotenv
openai==0.28